- Rename from utilies/command.ts to utilies/commands.ts - Minor cosmetic change to utilies/commands.ts
