+++ test/integration-tests/commands/build.test.ts
@@ -23,7 +23,7 @@ import { FolderContext } from "../../../src/FolderContext";
 import { WorkspaceContext } from "../../../src/WorkspaceContext";
 import { Commands } from "../../../src/commands";
 import { makeDebugConfigurations } from "../../../src/debugger/launch";
-import { Workbench } from "../../../src/utilities/command";
+import { Workbench } from "../../../src/utilities/commands";
 import { continueSession, waitForDebugAdapterCommand } from "../../utilities/debug";
 
 suite("Build Commands", function () {
